Remake inheritance to fix outgoing and incoming APIs
