+++ protocol/build.gradle.kts
@@ -30,4 +30,5 @@ dependencies {
 
     // Test dependencies
     testImplementation(libs.junit.jupiter)
+    testImplementation(libs.slf4j.simple)
 }
