+++ protocol/src/main/java/org/geysermc/mcprotocollib/network/packet/FakeFlushPacket.java
@@ -1,0 +1,5 @@
+package org.geysermc.mcprotocollib.network.packet;
+
+public record FakeFlushPacket() implements Packet {
+    public static final FakeFlushPacket INSTANCE = new FakeFlushPacket();
+}
