+++ protocol/src/main/java/org/geysermc/mcprotocollib/protocol/ServerListener.java
@@ -101,14 +101,14 @@ public class ServerListener extends SessionAdapter {
                 if (packet instanceof ClientIntentionPacket intentionPacket) {
                     switch (intentionPacket.getIntent()) {
                         case STATUS -> {
-                            protocol.setOutboundState(ProtocolState.STATUS);
-                            protocol.setInboundState(ProtocolState.STATUS);
+                            session.switchOutboundProtocol(() -> protocol.setOutboundState(ProtocolState.STATUS));
+                            session.switchInboundProtocol(() -> protocol.setInboundState(ProtocolState.STATUS));
                         }
                         case TRANSFER -> {
                             if (session.getFlag(MinecraftConstants.ACCEPT_TRANSFERS_KEY, false)) {
                                 beginLogin(session, protocol, intentionPacket, true);
                             } else {
-                                protocol.setOutboundState(ProtocolState.LOGIN);
+                                session.switchOutboundProtocol(() -> protocol.setOutboundState(ProtocolState.LOGIN));
                                 session.disconnect("Server does not accept transfers.");
                             }
                         }
@@ -140,8 +140,8 @@ public class ServerListener extends SessionAdapter {
                     session.enableEncryption(protocol.enableEncryption(key));
                     new Thread(() -> authenticate(session, key)).start();
                 } else if (packet instanceof ServerboundLoginAcknowledgedPacket) {
-                    protocol.setOutboundState(ProtocolState.CONFIGURATION);
-                    protocol.setInboundState(ProtocolState.CONFIGURATION);
+                    session.switchOutboundProtocol(() -> protocol.setOutboundState(ProtocolState.CONFIGURATION));
+                    session.switchInboundProtocol(() -> protocol.setInboundState(ProtocolState.CONFIGURATION));
 
                     // Credit ViaVersion: https://github.com/ViaVersion/ViaVersion/blob/dev/common/src/main/java/com/viaversion/viaversion/protocols/protocol1_20_5to1_20_3/rewriter/EntityPacketRewriter1_20_5.java
                     for (Map.Entry<String, Object> entry : networkCodec.entrySet()) {
@@ -195,14 +195,14 @@ public class ServerListener extends SessionAdapter {
                     // The developer who sends ClientboundStartConfigurationPacket needs to setOutboundState to CONFIGURATION
                     // after sending the packet. We can't do it in this class because it needs to be a method call right after it was sent.
                     // Using nettys event loop to change outgoing state may cause differences to vanilla.
-                    protocol.setInboundState(ProtocolState.CONFIGURATION);
+                    session.switchInboundProtocol(() -> protocol.setInboundState(ProtocolState.CONFIGURATION));
                 } else if (packet instanceof ServerboundPingRequestPacket pingRequestPacket) {
                     session.send(new ClientboundPongResponsePacket(pingRequestPacket.getPingTime()));
                 }
             }
             case CONFIGURATION -> {
                 if (packet instanceof ServerboundFinishConfigurationPacket) {
-                    protocol.setOutboundState(ProtocolState.GAME);
+                    session.switchOutboundProtocol(() -> protocol.setOutboundState(ProtocolState.GAME));
                     ServerLoginHandler handler = session.getFlag(MinecraftConstants.SERVER_LOGIN_HANDLER_KEY);
                     if (handler != null) {
                         handler.loggedIn(session);
@@ -218,13 +218,13 @@ public class ServerListener extends SessionAdapter {
 
     private void beginLogin(Session session, MinecraftProtocol protocol, ClientIntentionPacket packet, boolean transferred) {
         isTransfer = transferred;
-        protocol.setOutboundState(ProtocolState.LOGIN);
+        session.switchOutboundProtocol(() -> protocol.setOutboundState(ProtocolState.LOGIN));
         if (packet.getProtocolVersion() > protocol.getCodec().getProtocolVersion()) {
             session.disconnect(Component.translatable("multiplayer.disconnect.incompatible", protocol.getCodec().getMinecraftVersion()));
         } else if (packet.getProtocolVersion() < protocol.getCodec().getProtocolVersion()) {
             session.disconnect(Component.translatable("multiplayer.disconnect.outdated_client", protocol.getCodec().getMinecraftVersion()));
         } else {
-            protocol.setInboundState(ProtocolState.LOGIN);
+            session.switchInboundProtocol(() -> protocol.setInboundState(ProtocolState.LOGIN));
         }
     }
