+++ protocol/src/main/java/org/geysermc/mcprotocollib/protocol/codec/MinecraftCodec.java
@@ -228,10 +228,10 @@ public class MinecraftCodec {
             .protocolVersion(767)
             .helper(() -> new MinecraftCodecHelper(LEVEL_EVENTS, SOUND_NAMES))
             .minecraftVersion("1.21")
-            .state(ProtocolState.HANDSHAKE, PacketStateCodec.builder()
+            .state(ProtocolState.HANDSHAKE, PacketRegistryBuilder.builder()
                     .registerServerboundPacket(ClientIntentionPacket.class, ClientIntentionPacket::new)
             )
-            .state(ProtocolState.LOGIN, PacketStateCodec.builder()
+            .state(ProtocolState.LOGIN, PacketRegistryBuilder.builder()
                     .registerClientboundPacket(ClientboundLoginDisconnectPacket.class, ClientboundLoginDisconnectPacket::new)
                     .registerClientboundPacket(ClientboundHelloPacket.class, ClientboundHelloPacket::new)
                     .registerClientboundPacket(ClientboundGameProfilePacket.class, ClientboundGameProfilePacket::new)
@@ -243,12 +243,12 @@ public class MinecraftCodec {
                     .registerServerboundPacket(ServerboundCustomQueryAnswerPacket.class, ServerboundCustomQueryAnswerPacket::new)
                     .registerServerboundPacket(ServerboundLoginAcknowledgedPacket.class, ServerboundLoginAcknowledgedPacket::new)
                     .registerServerboundPacket(ServerboundCookieResponsePacket.class, ServerboundCookieResponsePacket::new)
-            ).state(ProtocolState.STATUS, PacketStateCodec.builder()
+            ).state(ProtocolState.STATUS, PacketRegistryBuilder.builder()
                     .registerClientboundPacket(ClientboundStatusResponsePacket.class, ClientboundStatusResponsePacket::new)
                     .registerClientboundPacket(ClientboundPongResponsePacket.class, ClientboundPongResponsePacket::new)
                     .registerServerboundPacket(ServerboundStatusRequestPacket.class, ServerboundStatusRequestPacket::new)
                     .registerServerboundPacket(ServerboundPingRequestPacket.class, ServerboundPingRequestPacket::new)
-            ).state(ProtocolState.CONFIGURATION, PacketStateCodec.builder()
+            ).state(ProtocolState.CONFIGURATION, PacketRegistryBuilder.builder()
                     .registerClientboundPacket(ClientboundCookieRequestPacket.class, ClientboundCookieRequestPacket::new)
                     .registerClientboundPacket(ClientboundCustomPayloadPacket.class, ClientboundCustomPayloadPacket::new)
                     .registerClientboundPacket(ClientboundDisconnectPacket.class, ClientboundDisconnectPacket::new)
@@ -274,7 +274,7 @@ public class MinecraftCodec {
                     .registerServerboundPacket(ServerboundPongPacket.class, ServerboundPongPacket::new)
                     .registerServerboundPacket(ServerboundResourcePackPacket.class, ServerboundResourcePackPacket::new)
                     .registerServerboundPacket(ServerboundSelectKnownPacks.class, ServerboundSelectKnownPacks::new)
-            ).state(ProtocolState.GAME, PacketStateCodec.builder()
+            ).state(ProtocolState.GAME, PacketRegistryBuilder.builder()
                     .registerClientboundPacket(ClientboundDelimiterPacket.class, ClientboundDelimiterPacket::new)
                     .registerClientboundPacket(ClientboundAddEntityPacket.class, ClientboundAddEntityPacket::new)
                     .registerClientboundPacket(ClientboundAddExperienceOrbPacket.class, ClientboundAddExperienceOrbPacket::new)
